Moved sound handling to utilsound.h/cpp
