+++ Client/qtTeamTalk/mainwindow.h
@@ -33,6 +33,7 @@
 #include <QNetworkAccessManager>
 
 #include "common.h"
+#include "utilsound.h"
 
 #if defined(Q_OS_WIN32)
 #define USE_POLL 0
