+++ Client/qtTeamTalk/textmessagedlg.cpp
@@ -24,6 +24,7 @@
 #include "textmessagedlg.h"
 #include "appinfo.h"
 #include "settings.h"
+#include "utilsound.h"
 #include <QDebug>
 
 #define LOCAL_TYPING_DELAY  5000
