+++ Client/qtTeamTalk/uservolumedlg.cpp
@@ -23,7 +23,7 @@
 
 #include "uservolumedlg.h"
 #include "appinfo.h"
-#include "common.h"
+#include "utilsound.h"
 
 #include <QMessageBox>
